requirements.txt
# This project uses Python 3.7

opencv-python==4.11.0.86
SoccerNet==0.1.62
pandas-1.3.5